Add a flags variable to the API for future proofing.
This unsigned 64bit variable will be used to cope with unexpected
changes that need to be done without having to resort to creating
a new call.  I added it to everything, even when it doesn't appear
to make sense as we don't know what will be needed in the future.

Daniel first proposed this idea on the mailing list a while back
as it is used quite heavily in libVirt. Thanks Daniel for the
suggestion!

Signed-off-by: Tony Asleson <tasleson@redhat.com>