Remove numerous if/else checks in async. display code.
When retrieving the result of an async. operation the code was
doing conditional checks to determine type and then display it.
I moved the code to generate the data in the correct format into
the classes themselves so that this could be done via OOP.
Much cleaner and code doesn't change later when we introduce new types.

Signed-off-by: Tony Asleson <tasleson@redhat.com>