Make pool parameter optional when replicating a volume.
Some arrays have specific requirements for which pools you can use to clone or
copy to.  The SMI-S specification has not defined how clients can determine
these requirements.  As most SMI-S providers will pick a suitable destination
pool based on the type of replication, passing what pool to use for the
replication is not needed and causes a burden on the client for little gain.
When the user passes Null/None for the pool parameter the plug-in will
select the most appropriate pool to use for the replication.

Signed-off-by: Tony Asleson <tasleson@redhat.com>