lsmcli: Change check for determining if lsmd is running
Instead of checking to see if the process exists we will check to
see if any unix domain sockets are available and then check to see
if we can use one of them.  This allows us to distinguish between
the daemon not running and a URI that specifies a non-existent plug-in.

Signed-off-by: Tony Asleson <tasleson@redhat.com>