Inconsistency corrections between C and Python API
There were a number of cases where the C and Python API were different with
regards to job handling.  In all of these cases the C API had a job id whereas
the python did not.  The functions that had these differences really didn't
need the ability to do asynchronous operations and thus there were removed.

Signed-off-by: Tony Asleson <tasleson@redhat.com>