Explicitly set large numeric as long
Large numeric constants in the range of (2^32 - 2^64) will be treated
differently on different arches for older python interpreters.  On 32bit
it would be promoted to a long where on 64bit it would remain an int.  The
compiled .pyc and .pyo files would differ for the exact same python source
files.  Append a L to the numeric constant to force the variable type to be
of type long.  This is fixed on newer version of the python interpreter.

Signed-off-by: Tony Asleson <tasleson@redhat.com>