Fix return values and optimization
Note: This change is a combination of two different
patches from Ma Shimiao with some modification made
by myself.

When we have values that can be lists of things it is
important that when that list is empty that we return
an empty vector type instead of a null type.

An example:

std::vector<Value> rc;
return rc;

and

return Value();

Return very different things when they get serialized to JSON.

Based on work by Ma Shimiao <mashimiao.fnst@cn.fujitsu.com>

Signed-off-by: Tony Asleson <tasleson@redhat.com>