ontap: Fix timeouts again.
The low level socket timeout was being thrown form the underlying
library, this is not being caught and translated into a FilterError.

The code in the exception handler for the resize was trying to
resize back to an absolute size, which isn't how NetApp API works.
It requires a delta resize, this caused the resize to fail.

Signed-off-by: Tony Asleson <tasleson@redhat.com>