ontap: Experimental disk listing functionality.
Support for the recently added disk functionality.

While adding support for this the following observations came to mind.

* We appear to be missing the disk method from the plug-in interface.

* The Disk member variables for STATUS, ENABLE_STATUS, and HEALTH seem a little
   confusing.  I had to reference the actual values to determine what exactly
   each was referring too.

   STATUS: Is a bit field and can have multiple values simultaneously.  Values
           can be things like: START, STOP, IN_SERVICE, COMPLETED, ERROR, OK

   ENABLE_STATUS: Single value with values such as: ENABLED, DISABLED, IN_TEST,
                  STARTING, QUIESCE

   HEALTH: Single value with values such as: OK, DEGRADED, MINOR_FAIL,
           MAJOR_FAIL, CRITICAL_FAIL, NON_RECOVERABLE_ERR.

Unfortunately, the NetApp API doesn't seem to cover STATUS & HEALTH.
ENABLE_STATUS is kind of covered as you can tell of a disk is enabled/disable.

I'm wondering if we shouldn't simplify some of the disk member variables to
least common denominator?

Signed-off-by: Tony Asleson <tasleson@redhat.com>