change Disk class for supporting optional properties
Took Gris's patch and modified it for the following:

* Create a properties class which is derived from IDATA

* Modifiy serilization code to also deserialize parameters
  before passing to the constructor

* Add an optional parameter to Disk class to take a Properties
  class.

The reason I am leaning this way is so that we can provide a new
data type with list/get/set functions in C too.

NOTES:

Want to implement the deserialization in C too to make sure
we can just as easily handle this change.

Signed-off-by: Tony Asleson <tasleson@redhat.com>