smis.py: Catch additional exceptions.
Added additional excepts in the exception decorator
handle_cim_errors.  One thing that became apparent in testing
is that we should not nest decorated methods.  Only the outer
most methods which are called and required by the framework
should have the @handle_cim_error decorator.  Otherwise, if
we have them nested we end up having the decorator run through
the exception logic more than once which can be a problem if a
raised exception is caught the second time through the
exception decorator.

Signed-off-by: Tony Asleson <tasleson@redhat.com>