lsmcli: Add ability to list plugins (V2)
Added the ability to list plugins.  This was requested by an end user and it
seems pretty useful, thus adding it.

The C plug-in API had to change a little to accommodate this as
previously you had to get a startup request before you filled in the description
and version for the plug-in.  This code works by opening the RPC and writing a
plugin_info command first before anything else is sent, thus the information is
not available.  As we are just trying to see what is available this is
problematic and was changed.

Notes:
* The API returns a list of strings.  Each string contains the description and
version separated by a user passed in value.  It would probably be cleaner
to create a new plug-in data type, but that seems a little overkill for such
a simple need.

* The python plug-in API differs slightly on how description and version are
supplied to the framework, this could be made the same.

* The smis plug-in proxy is going to break for eSeries if a user uses "interop"
for the namespace.  Need to make sure we clean that up or just merge the eSeries
changes back into the main smis plugin.

* C bits for the API need to be added to accomplish the same thing.  If folks
are OK with this patch I will add them.

Signed-off-by: Tony Asleson <tasleson@redhat.com>