common.py: Add decorator function for return types
Added the decorator function return_requires which can be used to ensure
that a given python function returns the required types.

This was added to the client API to ensure that the plug-in was providing
the correct return types.

Note: This only works when using the API over RPC.  If you run a python
plugin stand alone the decorators are not in use.

Signed-off-by: Tony Asleson <tasleson@redhat.com>