data.py: Change class attributes to private (V3)
This patch addresses the following:

1. The file data.py had some class variables which aliased python built-in
functionality (eg. id).  Makes it difficult when you actually need to use id.

2. Because of #1 when I changed the variables to not conflict we created a
problem where the classes didn't serialize to what was expected.  I changed
the built-in serialization to remove the '_' at the beginning of each attribute
when we convert to json, and to add it back when we serialize it to the python
class.

3. Lastly because all the class attributes are now prefixed with an '_' we now
have changed the interface where they are all private (good), but we need to
keep the same interface as before, so I created a decorator which can be used
to create the generic get/set property methods for each class.  Going forward as
the class needs updates we can drop the decorator for those attributes which
need more logic.  We could also update the decorator to add a callback for
value validation.

With these changes we better decouple the representation from the
presentation and what is serialized across the wire without changing external
use *.

* WARNING: The simulator data needs to be deleted (/tmp/lsm_sim_data) as the old
classes will have the old attributes stored in them, otherwise expect quick
failures with confusing exceptions.  I wouldn't think this would be a big issue,
but we need to add a version and version check in the simulator data to prevent
this from wreaking havoc on a user.

Signed-off-by: Tony Asleson <tasleson@redhat.com>