Use Argparse subcommands for commands
Put all info on commands into a "cmds" data structure, and then use that
to register subcommands for everything.

Set self.func to the function we should call based upon the convention
that the method is named the same as the command, but perhaps with '_'s
instead of '-'s. This required some methods to be renamed to match.

Signed-off-by: Andy Grover <agrover@redhat.com>