data.py: Re-factor and reduce code
I went through and consolidated functionality in the base class of IData and
all of the classes that derived from it.

The methods:

str_of_key
value_of_key
key_display_sequence

now have an default implementation.  I re-factored out the differences in the
implementations and added a method:

def value_convert(self, key_name, value, human, enum_as_number,
                      list_convert)

which is used to format a specific value for each of the specific classes.  It
has a default implementation which just returns value.

I also removed the unused methods:

column_headers
column_data

Code passes cmdtest.py, however...

NfsExport class is broken at the moment and needs fixes in cmdline.py.
Will address that with a subsequent patch.  We are missing NFS specific
testing in cmdtest.py, so I will add that too so that we can catch breakage
when it occurs.  NfsExport had a different display  compared to others, want to
avoid that again.

Signed-off-by: Tony Asleson <tasleson@redhat.com>