Differentiate subcommand-specific optional args from global optional args
subcommand-specific options are still lumped together with global options.
This patch extends the last patch and puts them in their own argument
group too.

Unfortunately there doesn't appear to be a way to add options
added via 'parents' to an argument group, so we're stuck with the name for
their group being 'optional arguments', unless we wanted to do something
more drastic, like not using 'parents' but adding all global options to
each subcommand ourselves.

In light of this, we try to make the two argument group names we *can* set
as specific as possible.

Also, add some whitespace for clarity.

Signed-off-by: Andy Grover <agrover@redhat.com>