udev: commented out UA rules except for REPORTED LUNS DATA HAS CHANGED
Testing has revealed an issue with multipath complaining about
the capacity changing on one path to a device, and discontinuing
the use of that path.  This is undesirable, as it appears that
at least one storage array does not always report the UA for
CAPACITY DATA HAS CHANGED when a LUN is resized.  Comment out
all the rules that rescan a device when attributes change until
this can be investigated more thoroughly.

Note that this is a preexisting issue, since anyone could have
rescanned a device via sysfs.  But the udev rules made this
automatic, so it could now occur all the time.

Leave the rule for REPORTED LUNS DATA HAS CHANGED in place, though.

Signed-off-by: Ewan D. Milne <emilne@redhat.com>