Python API: Reduce the visible public API
Given the convention of things that start with '_' are not to
be used, I went through the API and trimmed down the public
modules, functions, classes and their methods.  This was done
because once we declare the API good we cannot change the things
that are advertised as public.  This increases what we need to
maintain and what we are unable to re-factor and improve.

My goal is that whatever a user of the library or a plug-in
implementer will have everything they need access to by using
what is exposed in __init__.py.

No where should a plug-in or client use anything similar to:
   from lsm._common import Info

They should instead be doing:
   from lsm import Info

Changes made:
- A number of files were renamed with a leading '_'
- A number of items were removed from __init__.py, some added
- When importing things into files I choose to rename it so that it wouldn't
  be picked up as well eg. from abc import ABCMeta as _ABCMeta

At this point things are pretty clean, but there is one spot I would like to
review.  The classes  Pool & Disk has the following static methods:

    STATIC METHOD: disk_type_to_member_type
    STATIC METHOD: member_type_is_disk
    STATIC METHOD: member_type_to_disk_type
    STATIC METHOD: raid_type_to_str
    STATIC METHOD: thinp_type_str_to_type
    STATIC METHOD: thinp_type_to_str
    STATIC METHOD: disk_type_str_to_type
    STATIC METHOD: disk_type_to_str
    STATIC METHOD: status_to_str

These methods are either being used by the smi-s plug-in or the
simulator.  If these are absolutely required, fine.  However, lets
make sure that they are really needed in the general API or if they
should be put into the respective consumers logic.

Comments/Suggestions?

Note: I'm using a utility I wrote to dump what is public.  Once that
is polished I will post that too.

Thanks,
Tony

Signed-off-by: Tony Asleson <tasleson@redhat.com>