public_symbols.py: Simple tool for dumping visibility
Using python introspection this tool simple walks
the available symbols and dumps anything out that
doesn't start with a '_'.

It isn't perfect, for example:

class f(object):
    def __init__(self):
        self.a
        self.b

The members a & b will only show up if you
instantiate an instance of the class and then
run dir() on it.  Thus not all symbols can be
dumped with this tool as we are not instantiating
instances of every class.

Signed-off-by: Tony Asleson <tasleson@redhat.com>