C API: Clean up/reduce #includes
A number of header files included things they didn't need or already had.
In addition I changed the signature of the register function in C to pass the
URI as a string instead of a xmlURIPtr.  I did this so that a plug-in writer
could choose to use whatever they want to parse their own URI string.  It
also removes the requirement for the plug-in developer to have the libxml
development libraries.

I will be adding a library function that will parse the uri character string
into its parts, so that developers don't have to write their own.  The
implementation will not expose in the interface.

Signed-off-by: Tony Asleson <tasleson@redhat.com>