_data.py: Improve python json performance for RPC
The built in python json functionality is slower than simplejson, especially
when running on python included in EL6.  F20 the time difference isn't as
dramatic, just a second or two.

With this change, retrieving a list of 50K volumes takes ~3.2 seconds
instead of ~13.5 seconds when simplejson is available.  Previous tests I was
seeing ~25 seconds, not sure why I was getting those higher timings.

Notes:
- Python simulator was used with a python client that simply retrieved all
  volumes
- When loading the sim plug-in some time is used to read all 50K volumes from
  disk this time is not included in the above numbers
- Times above were recorded running on EL6
- simplejson returns values of 'str' instead of 'unicode' when consuming json
  from C simulator, haven't looked into why.

Client code used for test:

   import lsm
   import time

   c = lsm.Client('sim://')

   start = time.time()
   volumes = c.volumes()
   end = time.time()

   print '#Volumes', str(len(volumes)), 'in %f s' % (end-start)

Signed-off-by: Tony Asleson <tasleson@redhat.com>