_iplugin.py: Remove abstract base class from some classes
In _iplugin.py my original intention was that we specify the interface and then
plug-in writers provide the implementation.  However, in those cases where the
plug-in doesn't support one or more operations the plug-in writer was required
to provide an implementation which just raised an exception stating that the
operation wasn't supported.  Instead of requiring every plug-in to do this I
changed the classes so that the classes that have an interface that isn't
mandatory to provide this default not implemented implementation.

This allows for a reduction in the overall code base without loss of
functionality.  The base IPlugin class is still abstract thus the class that
derives from it will still need to implement all the methods contained in it.

Signed-off-by: Tony Asleson <tasleson@redhat.com>