lsmd: Fix clean_up function
The call to waitid was used incorrectly.  The code
incorrectly used the behavior of waitpid.  With this
change we will correctly cleanup all of the outstanding
exited child processes instead of just cleaning up one
process per call into this function.  With this change
we get <defunct> processes cleaned up much cleaner and
it's now working as initially intended.

Signed-off-by: Tony Asleson <tasleson@redhat.com>