C API & PY: Add plugin_data/optional_data to Pools V2
Python had optional_data, but C did not.

Note: The optional_data for python was retaining the value
type.  The C version requires that the value is always a
string, so I needed to put in changes to convert everything
to a string and convert things to numeric values when needed.

If we require this ability we will should probably store the
value in C with some type of variant structure and provide
functions to determine what type it is and then have a function
that will return it in that form.

V2: Remove debug code in simulator plugin

Signed-off-by: Tony Asleson <tasleson@redhat.com>