C API: optional data, add more value types
The existing code assumes that optional data is a key/value where the key is
a string type.  However, the python code is placing numeric values, string lists
and strings into the optional data which automatically gets serialized and
de-serialized back to the client retaining the types enclosed.  I made a change
in a previous commit which changes it to make everything a string and we broke
the display code which was expecting the member id of the optional pool data to
be a list of strings.

This patch adds the following capabilities for optional data type values for
the C API:

-Get/Set long double
-Get/Set uint64_t
-Get/Set int64_t
-Get/Set lsm_string_list

I added a enum and function:

typedef enum {
    LSM_OPTIONAL_DATA_INVALID       = -2,   /**< Invalid */
    LSM_OPTIONAL_DATA_NOT_FOUND     = -1,   /**< Key not found */
    LSM_OPTIONAL_DATA_STRING        = 1,    /**< Contains a string */
    LSM_OPTIONAL_DATA_SIGN_INT      = 2,    /**< Contains a signed int */
    LSM_OPTIONAL_DATA_UNSIGNED_INT  = 3,    /**< Contains an unsigned int */
    LSM_OPTIONAL_DATA_REAL          = 4,    /**< Contains a real number */
    LSM_OPTIONAL_DATA_STRING_LIST   = 10    /**< Contains a list of strings*/
} lsm_optional_data_type;

lsm_optional_data_type LSM_DLL_EXPORT lsm_optional_data_type_get(
                                    lsm_optional_data *op, const char *key);

Which needs to be utilized to see that type the value is.  Once that is
determined then the user can call one of the newly added corresponding get
functions.

However, this is far from ideal for the following reason.  The JSON RPC IPC
used between the client and plug-in only has a single number type which
encompasses all of the types listed above (int64_t, uint64_t, long double)
and then some (numbers of arbitrary size).  Thus a plug-in could be returning
a value to a client and the 'type' of numeric may vary based on the numerical
value.  So a client to be robust would need to  handle checking for multiple
numeric types for a single optional value. The only way I see this being
consistent is if we extend the API so that we encode the 'type' and value in
the serialized JSON so that we can consistently give the same 'type' of numeric
to the client application.  This patch looks at the numeric string and puts it
in the first available fitting number in this order: int64, uint64, long double.

Any other suggestions people have, please share them.

Thanks!

Regards,
Tony

Signed-off-by: Tony Asleson <tasleson@redhat.com>