Changes to get unit test to work
Gris asked me to remove the patch
'lsmcli: Remove initiator commands and trivial changes'.

However, that patch did have some critial changes to make things
work so parts of it that are needed have been added to this one
and the changes needed to get the unit tests to pass.

Signed-off-by: Tony Asleson <tasleson@redhat.com>