C & PY API: Remove optional data
After discussion on the mailing list it came to light that the
whole notion of optional data is more of a hinderence than a
benefit.  This library needs to be easy and straight forward
to use.

I renamed the C lsm_optional_data to lsm_hash as we are using it
for parsing uri and use the hash for query string parameters.

"No matter how far down the wrong road you’ve gone, turn back."

Signed-off-by: Tony Asleson <tasleson@redhat.com>