PY: Rethrowing existing exceptions
When you have:

try:
    foo()
except Exception as e:
    # Want to raise original exception
    raise  # 1. preferred

    # raise e # 1a.  Bad

try:
    bar()
except Exception as e:
    import sys
    exception_info = sys.exec_info()

    try:
        bar()
    except Exception as e_inner:
        pass

    raise exception_info[1], None, exception_info[2]

It's better to do #1 instead of #1a as #1 will preserve the
stack trace to where the actual error occurred instead of
leading back to #1a  where it was simply re-thrown.

However, in some cases you need to do a bit of work that may throw
an exception itself when handling an exception, so you need to
preserve and rethrow.  In this case follow #2.

See:
http://nedbatchelder.com/blog/200711/rethrowing_exceptions_in_python.html

Signed-off-by: Tony Asleson <tasleson@redhat.com>