Python plugins: Re-work generic exceptions
It wasn't very useful to just print the exception text without
the stack trace, so we are adding that too.

Note: The reason this is not in a common function is it's
possible that the common function could not be found in
the namespace and thus no debug information would be
available.  Try and keep it simple and always available.

Signed-off-by: Tony Asleson <tasleson@redhat.com>