Remove txt_a from _data.py
A more Pythonic way of generating a comma-delimited string of 0 to n
elements is to put them in a list and then ", ".join() them.

However, in the usage in smis.py, it doesn't actually appear the
status_info is actually being added-to more than once, so it's
possible status_info could just be a string that we set to the new value,
but I've kept the code as-is for now.

tasleson: Fixed syntax error in Andy's patch.

Signed-off-by: Andy Grover <agrover@redhat.com>
Signed-off-by: Tony Asleson <tasleson@redhat.com>