smis.py: Add debug for invoke request/reply
Added to aide in debugging providers.  For smis.py, if you add
debug_path to the query string anytime an invoke ends in error
we will dump the xml request/reply to a file in the directory
specified by 'debug_path'.

Note: renamed a variable as to not conflict with the os import.

Signed-off-by: Tony Asleson <tasleson@redhat.com>