tester.c:  Fix hang and FS space exhaustion
The issue is that when you go through teardown, if you use
fail_unless in it if you actuall fail you end up calling
teardown again and you get stuck in a loop until you exhaust
the fs with a tmpfile.

For this particular issue to show up you just need to cause an error
when closing the connection which is what happens when you close the
connection on a plug-in that exited already.

Gris, please verify it works for you too.

Signed-off-by: Tony Asleson <tasleson@redhat.com>