ontap.py: Use just logical unit name with no path
Using the full path and name for the name is an implementation
detail.  Changed to just using name and the plugin_data storing
the full path and name instead.

Also fixed a bug where the pool id for the volumes was incorrect.  It
always showed the id for the aggregate that the NetApp volume was on.
Fixed this and I also changed so we wouldn't display the NetApp volume
that gets created if a user decides to create a logical unit on an aggregate
instead of selecting a pool that is a NetApp volume (fs).  This was done so
that when the user passes the base aggregate, the newly created lsm volume
pool id would match the one passed in, not some NetApp volume that was
created as a container to house the newly created logical volume.

Signed-off-by: Tony Asleson <tasleson@redhat.com>