ontap.py: Remove the 'hidden' NetApp Volume (lun container)
Require the user to use a pool that is technically a NetApp volume which we show
call a FS when creating a logical unit.  This removes a lot of complexity
in creating/resizing and maintaining a NetApp volume for the sole purpose
of being able to create a volume/lun.

Signed-off-by: Tony Asleson <tasleson@redhat.com>