Replace NO_MAPPING with NO_STATE_CHANGE V2
 * Remove lsm.ErrorNumber.NO_MAPPING and LSM_ERR_NO_MAPPING.
 * Change volume_unmask() to use NO_STATE_CHANGE in stead of NO_MAPPING.

V2: lsm_plugin_ipc.cpp, the failed call to value_to_fs is likely a memory
allocation error.  In 9 other places we are returning LSM_ERR_NO_MEMORY so
we will be consistent and return the same here.  There is some ambiguity
as if the 'Value' we pass to the function isn't a FS object instance we
will return NULL too.  I will follow-up with a future patch to remove this
ambiguity in all the places where we convert a 'Value' to the appropriate
object C structure pointer.

Signed-off-by: Gris Ge <fge@redhat.com>
Signed-off-by: Tony Asleson <tasleson@redhat.com>