C: Remove ambiguity in class checking
The function value_to_* make sure that the value we are converting
too matches the class in question.  If for some reason the json
does not match expected we return a NULL, which is the same as if
we had a failure to allocate memory.  Added checks before to
ensure that we have the correct json representation of the class
before we convert from Value type to C struct type.

Signed-off-by: Tony Asleson <tasleson@redhat.com>