cmdline.py: Remove optional --init argument for volume-mask
Targetd is the only plug-in that doesn't support access groups.  We
tried to make this work by faking each initiator to look as a standalone
access group, however this has a number of issues and didn't work as
advertised.  As this feature never worked, it cannot have been used
by anyone.  We are choosing to remove it and add support for access groups
to targetd and the targetd plugin.

Notes:
* Access group id is now required, not optional
* The underlying API needs no changes
* The man page didn't document this (thus no update)
* Masking is indeed broken and continutes to be broken for targetd
  with patches in the works

Signed-off-by: Tony Asleson <tasleson@redhat.com>