nfs export anon_uid/gid clarification and corrections
The nfs export anonymous uid & gid are suppose to be integers.  The command
line was passing None which overrode the python default of -1.  This didn't
cause any issues as the plugins were only checking for existence, not value.
However, this did cause a problem if a user called the
API without specifying the uid & gid as the default integer value of -1 caused
the ontap plugin to fail with an exception.

Unfortunately the C API declares the uid & gid as uint64_t so the default value
of -1 gets interpreted as (2**64-1).  Thus python plugins should check against
-1 and 18446744073709551615L to see if the argument is == default.  They will
get -1 from python clients and 18446744073709551615L from C clients.

Signed-off-by: Tony Asleson <tasleson@redhat.com>