C API: Fix simc plugin and correct plugin header doxygen parameters
Plugin IPC for C plugins wasn't allowing some of the parameters to be NULL,
which they can be according to the header documentation.  Fixed up code to
accomodate this valie use case.  Also fixed a bug in ss_restore, we were
checking v_files twice instead of v_files and v_restore_file.

Signed-off-by: Tony Asleson <tasleson@redhat.com>