Fix code and spurious log message
In the cases where the command line is exiting because of an
invalid parameter we are not allowing the client library to
close cleanly causing the plugin to report a syslog message.

Fixed this error case and added a check to the plugin to only
report this message if the client is not following protocol.

Signed-off-by: Tony Asleson <tasleson@redhat.com>