lsm_daemon.c: Fix coverity warning
warning: Null pointer passed as an argument to a 'nonnull' parameter

char *path_form(const char* path, const char *name)
{
    size_t s = strlen(path) + strlen(name) + 2;

This was being caused because coverity tool doesn't appear to
recognize that our log function exits on error.  Thus it makes
the conclusion that we are indeed passing NULL as "name" to
the path_form function.  I restructured the code so that we do
the bulk of the operations in the 'checked' leg of the if
statement.

Signed-off-by: Tony Asleson <tasleson@redhat.com>