cmdline.py: Correctly exit with non-zero exit
With commit 059f652884f0c34c60ce3e6efc63bbbcba56ec09 an error was
introduced which caused argparse failures to exit with a good status.  This
was caused because argparse is raising a SystemExit exception which was not
being caught.  Thus the err_exit variable was not be set to something other
than non-zero and command line callers were incorrectly getting good status
when in fact the command had failed.

Added except legs for SystemExit and any exception to cover the argparse
and any future exception that could possibly occur.

ref: https://bugzilla.redhat.com/show_bug.cgi?id=1238737

Signed-off-by: Tony Asleson <tasleson@redhat.com>