targetd.py: Fix error paths
Replaced the exception handlers with a common one that maps errors from targetd
to error codes with libStorageMgmt with default messages if applicable.  This
way we don't need to have a bunch of if/else statements in the exception
handlers.

Note: I added a variable so if a caller wants to get the exception to evaluate
they can.

Signed-off-by: Tony Asleson <tasleson@redhat.com>