Use substring matching for HPSA system mode
Some HPSA cards may report their mode as "RAID Mode" instead of "RAID".
Feasibly, a card could also report "HBA Mode" in a similar fashion.
If we use substring matching to look for "RAID" or "HBA" in the mode
string, we can support "RAID Mode" in addition to any other variation
so long as the mode has the "RAID" text in it. Ditto for "HBA".