changing exception to NO STATE CHANGE
If no LEDs are blinking, the arcconf utility aborts the command and
this results in an exception in arcconf plugin and having exception
'PLUGIN_BUG' will result the test to fail. Ideally, we need to check
for 'isBlinking' status for each Drives belonging to the volume.
I will create an issue for the same to take it up later.