sim_array [volume|fs]_resize: Change re-size behavior
The simulator only allows size increments of a block size.  The code was
returning success if the user specified 1 byte more for the resize, but
internally it wasn't updating the size, eg. it was leaving the previous
rounded down size.  This goes against the documentation for the volume
re-size where we may round up for resize.  This change rounds up the
supplied value.  If the rounded up size matches current size we return
no state change, else we round up and update the simulator value to match.

This also matches the expected behavior that when successful the new size
will indeed be different, although it may be larger than specified and
if the exact operation is repeated, it will fail with no state change.

Signed-off-by: Tony Asleson <tasleson@redhat.com>