configure.ac: Auto use python3 if available
Fedora 28 was the last version that allows us to build and test
both python2 and python3 code bases.  Change configure to select
python3 if it's present or if a user specifies it on command line.

Otherwise we will fail to configure on newer distributions which
are python3 only and the user doesn't specifiy --with-python3.

Signed-off-by: Tony Asleson <tasleson@redhat.com>