Use yajl if available
The default is to utilize a c++11 compiler.  When compiling on
older systems you are required to specify --with-yajl to not
have this requirement.  Changing behavior so if the system has
the header files for yajl it will be used instead or can be
forced to be used by using --with-yajl.

Signed-off-by: Tony Asleson <tasleson@redhat.com>