cmdline: Speed up volume listings
When we display volumes we try to see if the volume matches a locally
available disk.  We can check about 1K disk a second, but when we have
50K+ volumes the command line slows way down.  Instead of repeatedly
checking what the paths might be for each and every volume entry that
has a vpd83, we will first query all local disk and try to retrieve
its vpd and stuff it in a hash.  Before we spend the time to call
vpd83_search we will consult this lookup first.

In testing, instead of taking ~60 seconds to display 50K+ volumes it now
takes ~3 seconds.

Signed-off-by: Tony Asleson <tasleson@redhat.com>