configure.ac: Skip check with --without-smispy
When a user specifies that they do not want smispy, do not
include it in the `make check` test.  We do this because
they likely don't have pywbem, thus the check will fail.

The other plugins don't have external dependencies which would
prevent them from starting up when we list all the plugins.

Signed-off-by: Tony Asleson <tasleson@redhat.com>