cmdline.py: Requre subparser
At some point argparse added a 'required' option for sub-parsers
that defaults to 'False'.  When this happened the command line
breaks as apparently we need the sub-parser to evaluate for the values
set with 'set-defaults()'.  Otherwise we can get errors like:

$ lsmcli -u sim://
Traceback (most recent call last):
  File "/home/tasleson/projects/libstoragemgmt/python_binding/lsm/lsmcli/cmdline.py", line 69, in cmd_line_wrapper
    cli = CmdLine()
  File "/home/tasleson/projects/libstoragemgmt/python_binding/lsm/lsmcli/cmdline.py", line 2015, in __init__
    self.tmo = int(self.args.wait)
TypeError: int() argument must be a string, a bytes-like object or a real number, not 'NoneType'

in addition to not requiring a 'command' which we do.

Signed-off-by: Tony Asleson <tasleson@redhat.com>