cmdline.py: Add --fork_plugin debug dev. option
When developing a new plugin it's helpful to not require the daemon to be
running.  Add a hidden argument that when specified allows the command line
to exec. the plugin directly.  Historically python plugins were easier
to develop as you can execute a python plugin directly from the command
line.  With this functionality it allows plugin developers an easier
time developing in other languages too.

Signed-off-by: Tony Asleson <tasleson@redhat.com>