coverity: Remove toctou error
Coverity is reporting the following.

Error: TOCTOU (CWE-367): [#def2]
scan-scsi-target.c:120: fs_check_call: Calling function "stat" to perform check on "sysfs_path".
scan-scsi-target.c:213: toctou: Calling function "open" that uses "sysfs_path" after a check
function. This can cause a time-of-check, time-of-use race condition.
118|       strcat(sysfs_path, devpath);
119|
120|->     if (stat(sysfs_path, &sysfs_stat) < 0) {
121|           fprintf(stderr, "Cannot stat '%s': %s\n", sysfs_path, strerror(errno));
122|           usage(argv, 1);

This is a false positive because the recycled variable does not refer
to the same path.  Remove this by introducing a new variable for the
directory check.

Signed-off-by: Tony Asleson <tasleson@redhat.com>